Lift and shift combat calculation code into more cromulent files
