+++ src/CommandingOfficers/Commander.java
@@ -18,9 +18,9 @@ import CommandingOfficers.Modifiers.COModifier;
 import Engine.GameAction;
 import Engine.GameScenario;
 import Engine.XYCoord;
-import Engine.Combat.BattleInstance.BattleParams;
-import Engine.Combat.BattleInstance.CombatContext;
+import Engine.Combat.BattleParams;
 import Engine.Combat.BattleSummary;
+import Engine.Combat.CombatContext;
 import Engine.GameEvents.GameEventListener;
 import Engine.GameEvents.GameEventQueue;
 import Terrain.GameMap;
