+++ src/CommandingOfficers/Venge.java
@@ -6,9 +6,9 @@ import CommandingOfficers.Modifiers.CODamageModifier;
 import CommandingOfficers.Modifiers.CODefenseModifier;
 import CommandingOfficers.Modifiers.COModifier;
 import Engine.GameScenario;
-import Engine.Combat.BattleInstance.BattleParams;
-import Engine.Combat.BattleInstance.CombatContext;
+import Engine.Combat.BattleParams;
 import Engine.Combat.BattleSummary;
+import Engine.Combat.CombatContext;
 import Engine.GameEvents.GameEventQueue;
 import Terrain.MapMaster;
 import Units.Unit;
