+++ src/Engine/Combat/CombatEngine.java
@@ -1,9 +1,20 @@
 package Engine.Combat;
 
+import java.util.AbstractMap;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Map.Entry;
+
 import Terrain.GameMap;
 import Terrain.MapMaster;
 import Units.Unit;
+import Units.WeaponModel;
 
+/**
+ * CombatEngine provides COs with two levels of control of what goes on in combat:
+ * changeCombatContext(), which takes in a CombatContext and allows the COs to fiddle with just about anything.
+ * applyCombatModifiers(), which is for directly modifying a unit's firepower in a given attack.
+ */
 public class CombatEngine
 {
   /**
@@ -13,10 +24,7 @@ public class CombatEngine
    */
   public static BattleSummary calculateBattleResults( Unit attacker, Unit defender, MapMaster map, int attackerX, int attackerY )
   {
-    // Set up our combat scenario.
-    BattleInstance params = new BattleInstance(attacker, defender, map, attackerX, attackerY, false);
-
-    return params.calculateBattleResults();
+    return calculateBattleResults(attacker, defender, map, attackerX, attackerY, false);
   }
 
   /**
@@ -25,10 +33,76 @@ public class CombatEngine
    */
   public static BattleSummary simulateBattleResults( Unit attacker, Unit defender, GameMap map, int attackerX, int attackerY )
   {
-    // Set up our combat scenario.
-    BattleInstance params = new BattleInstance(attacker, defender, map, attackerX, attackerY, true);
-
-    return params.calculateBattleResults();
+    return calculateBattleResults(attacker, defender, map, attackerX, attackerY, true);
   }
 
+  /**
+   * Calculate and return the results of a battle.
+   * This will not actually apply the damage taken; this is done later in BattleEvent.
+   * Note that we assume the defender is where he thinks he is, but the attacker may move before attacking, so we take that coordinate explicitly.
+   * @param isSim Determines whether to cap damage at the HP of the victim in question
+   * @return A BattleSummary object containing all relevant details from this combat instance.
+   */
+  public static BattleSummary calculateBattleResults(Unit attacker, Unit defender, GameMap map, int attackerX, int attackerY, boolean isSim)
+  {
+    int defenderX = defender.x; // This variable is technically not necessary, but provides consistent names with attackerX/Y.
+    int defenderY = defender.y;
+    boolean attackerMoved = attacker.x != attackerX || attacker.y != attackerY;
+    int battleRange = Math.abs(attackerX - defenderX) + Math.abs(attackerY - defenderY);
+    WeaponModel attackerWeapon = attacker.chooseWeapon(defender.model, battleRange, attackerMoved);
+    WeaponModel defenderWeapon = null;
+    // Only attacks at point-blank range can be countered
+    if( 1 == battleRange )
+      defenderWeapon = defender.chooseWeapon(attacker.model, battleRange, false);
+
+    CombatContext context = new CombatContext(map, attacker, attackerWeapon, defender, defenderWeapon, battleRange, attackerX, attackerY);
+    
+    // If the attacker and defender get swapped, we want to still give a coherent BattleSummary.
+    Map<Unit, Entry<WeaponModel,Double>> unitDamageMap = new HashMap<Unit, Entry<WeaponModel,Double>>();
+    unitDamageMap.put(attacker, new AbstractMap.SimpleEntry<WeaponModel,Double>(attackerWeapon, 0.0));
+    unitDamageMap.put(defender, new AbstractMap.SimpleEntry<WeaponModel,Double>(defenderWeapon, 0.0));
+    
+    // let the COs fool around with anything they want...
+    attacker.CO.changeCombatContext(context);
+    defender.CO.changeCombatContext(context);
+    
+    // From here on in, use context variables only (Easy way to confirm: delete BattleInstance's variable block and see what gets marked as an error)
+    
+    double attackerHPLoss = 0;
+
+    // Set up our scenario.
+    BattleParams attackInstance = BattleParams.getAttack(context);
+
+    // Last-minute adjustments.
+    context.attacker.CO.applyCombatModifiers(attackInstance, true);
+    context.defender.CO.applyCombatModifiers(attackInstance, false);
+
+    double defenderHPLoss = attackInstance.calculateDamage();
+    unitDamageMap.put(context.attacker, new AbstractMap.SimpleEntry<WeaponModel,Double>(context.attackerWeapon, defenderHPLoss));
+    if( !isSim && defenderHPLoss > context.defender.getPreciseHP() ) defenderHPLoss = context.defender.getPreciseHP();
+
+    // If the unit can counter, and wasn't killed in the initial volley, calculate return damage.
+    if( context.canCounter && (context.defender.getPreciseHP() > defenderHPLoss) )
+    {
+      // New battle instance with defender counter-attacking.
+      BattleParams defendInstance = BattleParams.getCounterAttack(context);
+      defendInstance.attackerHP = Math.ceil(context.defender.getPreciseHP() - defenderHPLoss); // Account for the first attack's damage to the now-attacker.
+
+      // Modifications apply "attacker first", and the defender is now the attacker.
+      context.defender.CO.applyCombatModifiers(defendInstance, true);
+      context.attacker.CO.applyCombatModifiers(defendInstance, false);
+
+      attackerHPLoss = defendInstance.calculateDamage();
+      unitDamageMap.put(context.defender, new AbstractMap.SimpleEntry<WeaponModel,Double>(context.defenderWeapon, attackerHPLoss));
+      if( !isSim && attackerHPLoss > context.attacker.getPreciseHP() ) attackerHPLoss = context.attacker.getPreciseHP();
+    }
+    
+    // Calculations complete. We are setting up our BattleSummary, so go back to using known-accurate info
+    return new BattleSummary(attacker, unitDamageMap.get(attacker).getKey(),
+                             defender, unitDamageMap.get(defender).getKey(), 
+                             map.getEnvironment(attackerX, attackerY).terrainType,
+                             map.getEnvironment(defenderX, defenderY).terrainType, 
+                             unitDamageMap.get(defender).getValue(), unitDamageMap.get(attacker).getValue());
+  }
+  
 }
